♻️refactor: llm 패키지 통합
